new file:   src/Control/Kontrol.py 	modified:   src/GUI/GUI.py
